+++ src/evaluation_system/__init__.py
@@ -38,7 +38,7 @@ warnings.filterwarnings(
 )
 
 
-__version__ = "2406.0.0"
+__version__ = "2406.0.1"
 
 if __name__ == "__main__":
     print(__version__)