+++ docs/source/whats-new.rst
@@ -8,6 +8,12 @@ What's new
 v2406.0.0
 ~~~~~~~~~
 
+New Features
+++++++++++++
+- The history method and cli history command can expand the query
+  to all users with ``all_users=True`` in the python module or
+  ``--all-users`` in the cli
+
 Bug fixes
 +++++++++
 
@@ -112,7 +118,7 @@ New Features
 ++++++++++++
 - User data can be ingested also if not all data files match directory reference
   standard.
-- A new databrowser key (uri) was added that representing the object storage
+- A new databrowser key (uri) was added which represents the object storage
   where the data is stored.
 
 Breaking changes